[CORE] Js: simplify while false + dead code simplification + tests refactor
